eme: check statuses as soon as the MediaKeySession is listened to
