+++ src/core/eme/check_key_statuses.ts
@@ -1,0 +1,94 @@
+/**
+ * Copyright 2015 CANAL+ Group
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { ICustomMediaKeySession } from "../../compat";
+import { EncryptedMediaError } from "../../errors";
+import {
+  IEMEWarningEvent,
+  IKeySystemOption,
+} from "./types";
+
+const KEY_STATUSES = { EXPIRED: "expired",
+                       INTERNAL_ERROR: "internal-error",
+                       OUTPUT_RESTRICTED: "output-restricted" };
+
+/**
+ * Look at the current key statuses in the sessions and:
+ *   - construct the appropriate warnings
+ *   - returns blacklistedKeyIDs
+ *
+ * Throws if one of the keyID is on an error for which we should not fallback.
+ * @param {MediaKeySession} session - The MediaKeySession from which the keys
+ * will be checked.
+ * @param {Object} keySystem - Configuration. Used to known on which situations
+ * we can fallback.
+ * @returns {Array} - First element are the warnings to send.
+ * The second are keyIDs to blacklist.
+ */
+export default function checkKeyStatuses(
+  session : MediaKeySession | ICustomMediaKeySession,
+  keySystem: IKeySystemOption
+) : [IEMEWarningEvent[], ArrayBuffer[]] {
+  const { fallbackOn = {} } = keySystem;
+  const warnings : IEMEWarningEvent[] = [];
+  const blacklistedKeyIDs : ArrayBuffer[] = [];
+
+  (session.keyStatuses as any).forEach((_arg1 : unknown, _arg2 : unknown) => {
+    const [keyStatus, keyId] = (() => {
+      return (typeof _arg1  === "string" ? [_arg1, _arg2] :
+                                           [_arg2, _arg1]
+             ) as [MediaKeyStatus, ArrayBuffer];
+    })();
+
+    switch (keyStatus) {
+      case KEY_STATUSES.EXPIRED: {
+        const error = new EncryptedMediaError("KEY_STATUS_CHANGE_ERROR",
+                                              "A decryption key expired");
+
+        if (keySystem.throwOnLicenseExpiration !== false) {
+          throw error;
+        }
+        warnings.push({ type: "warning", value: error });
+        break;
+      }
+
+      case KEY_STATUSES.INTERNAL_ERROR: {
+        const error = new EncryptedMediaError("KEY_STATUS_CHANGE_ERROR",
+                                              "An invalid key status has been " +
+                                              "encountered: " + keyStatus);
+        if (fallbackOn.keyInternalError !== true) {
+          throw error;
+        }
+        warnings.push({ type: "warning", value: error });
+        blacklistedKeyIDs.push(keyId);
+        break;
+      }
+
+      case KEY_STATUSES.OUTPUT_RESTRICTED: {
+        const error = new EncryptedMediaError("KEY_STATUS_CHANGE_ERROR",
+                                              "An invalid key status has been " +
+                                              "encountered: " + keyStatus);
+        if (fallbackOn.keyOutputRestricted !== true) {
+          throw error;
+        }
+        warnings.push({ type: "warning", value: error });
+        blacklistedKeyIDs.push(keyId);
+        break;
+      }
+    }
+  });
+  return [warnings, blacklistedKeyIDs];
+}
